Small adaptation for gateway resources
