+++ src/main/kotlin/org/radarcns/gateway/io/DecompressInterceptor.kt
@@ -3,8 +3,6 @@ package org.radarcns.gateway.io
 import org.radarbase.io.lzfse.LZFSEInputStream
 import org.radarcns.gateway.Config
 import org.radarcns.gateway.inject.ProcessAvro
-import java.io.IOException
-import java.io.InputStream
 import java.util.zip.GZIPInputStream
 import javax.annotation.Priority
 import javax.inject.Singleton
@@ -13,7 +11,6 @@ import javax.ws.rs.core.Context
 import javax.ws.rs.ext.Provider
 import javax.ws.rs.ext.ReaderInterceptor
 import javax.ws.rs.ext.ReaderInterceptorContext
-import kotlin.math.min
 
 /**
  * Decompresses GZIP-compressed data. Uncompressed data is not modified. The data is decompressed
@@ -23,7 +20,7 @@ import kotlin.math.min
 @ProcessAvro
 @Singleton
 @Priority(Priorities.ENTITY_CODER)
-class GzipDecompressInterceptor(@Context private val config: Config) : ReaderInterceptor {
+class DecompressInterceptor(@Context private val config: Config) : ReaderInterceptor {
     override fun aroundReadFrom(context: ReaderInterceptorContext): Any {
         val encoding = context.headers?.getFirst("Content-Encoding")?.toLowerCase()
         if (encoding == "gzip") {
@@ -33,29 +30,5 @@ class GzipDecompressInterceptor(@Context private val config: Config) : ReaderInt
         }
         context.inputStream = LimitedInputStream(context.inputStream, config.maxRequestSize)
         return context.proceed()
-    }
-
-    class LimitedInputStream(private val subStream: InputStream, private val limit: Long): InputStream() {
-        private var count: Long = 0
-        override fun available() = min(subStream.available(), (limit - count).toInt())
-
-        override fun read(): Int {
-            return subStream.read().also {
-                if (it != -1) count++
-                if (count > limit) throw IOException("Stream size exceeds limit $limit")
-            }
-        }
-
-        override fun read(b: ByteArray?, off: Int, len: Int): Int {
-            if (len > 0 && count == limit) throw IOException("Stream size exceeds limit $limit")
-            return subStream.read(b, off, min((limit - count).toInt(), len)).also {
-                if (it != -1) count += it
-                if (count > limit) throw IOException("Stream size exceeds limit $limit")
-            }
-        }
-
-        override fun close() {
-            subStream.close()
-        }
     }
 }
